Fixed MessageService, all blocks build in theirs services
