+++ src/main/java/com/community/tools/service/discord/DiscordService.java
@@ -1,10 +1,20 @@
 package com.community.tools.service.discord;
 
+import com.community.tools.model.Event;
+import com.community.tools.model.EventData;
 import com.community.tools.service.MessageService;
+import com.community.tools.service.PublishWeekStatsService;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.entities.TextChannel;
@@ -18,7 +28,7 @@ import org.springframework.stereotype.Service;
 @Data
 @RequiredArgsConstructor
 @Profile("discord")
-public class DiscordService implements MessageService {
+public class DiscordService implements MessageService<MessageEmbed> {
 
   private JDA jda;
 
@@ -92,6 +102,64 @@ public class DiscordService implements MessageService {
   public <T> void sendBlockMessageToConversation(String channelName, T message) {
     jda.getTextChannelById(getIdByChannelName(channelName))
         .sendMessage((MessageEmbed) message).queue();
+  }
+
+  @Override
+  public MessageEmbed nextTaskMessage(List<String> tasksList, int numberTask) {
+    return new EmbedBuilder()
+        .addField("", MessagesToDiscord.NEXT_TASK + tasksList.get(numberTask) + ") :link:", false)
+        .build();
+  }
+
+  @Override
+  public MessageEmbed ratingMessage(String url, String img) {
+    return  new EmbedBuilder()
+        .addField("","Рейтинг этой недели доступен по ссылке: ", false)
+        .addField("","[:loudspeaker: click_me_123](" + url + ")", false)
+        .addField("","[Image](" + img + ")", true)
+        .build();
+  }
+
+  @Override
+  public MessageEmbed statisticMessage(List<EventData> events) {
+
+    EmbedBuilder embedBuilder = new EmbedBuilder();
+
+    Map<String, List<EventData>> sortedMapGroupByActors = new HashMap<>();
+    events.stream().filter(ed -> !sortedMapGroupByActors.containsKey(ed.getActorLogin()))
+        .forEach(ed -> sortedMapGroupByActors.put(ed.getActorLogin(), new ArrayList<>()));
+
+    embedBuilder.addField("", "`Statistic:`", false);
+
+    events.stream()
+        .collect(Collectors.groupingBy(EventData::getType))
+        .entrySet().stream()
+        .sorted(Comparator
+            .comparingInt((Entry<Event, List<EventData>> entry)
+                -> entry.getValue().size()).reversed())
+        .forEach(entry -> {
+          entry.getValue().forEach(e -> sortedMapGroupByActors.get(e.getActorLogin()).add(e));
+
+          embedBuilder.addField("", PublishWeekStatsService.getTypeTitleBold(entry.getKey())
+              + PublishWeekStatsService.emojiGen(entry.getKey()) + ": "
+              + entry.getValue().size(), false);
+        });
+    embedBuilder.addField("", "`Activity:`", false);
+
+    sortedMapGroupByActors.entrySet().stream()
+        .sorted(Comparator
+            .comparingInt((Entry<String, List<EventData>> entry)
+                -> entry.getValue().size()).reversed())
+        .forEach(name -> {
+          StringBuilder authorsActivMessage = new StringBuilder();
+          name.getValue()
+              .forEach(eventData -> {
+                authorsActivMessage.append(PublishWeekStatsService.emojiGen(eventData.getType()));
+              });
+          embedBuilder.addField("", name.getKey() + ": "
+              + authorsActivMessage,false);
+        });
+    return embedBuilder.build();
   }
 
   /**
