+++ src/test/java/com/community/tools/service/PublishWeekStatsServiceTest.java
@@ -38,9 +38,6 @@ class PublishWeekStatsServiceTest {
   @Mock
   private MessageService messageService;
 
-  @Mock
-  private BlockService blockService;
-
   @BeforeAll
   public void initMocks() {
     MockitoAnnotations.initMocks(this);
@@ -69,7 +66,7 @@ class PublishWeekStatsServiceTest {
         new EventData(new Date(), "roman", Event.COMMENT));
 
     Mockito.when(ghEventService.getEvents(any(), any())).thenReturn(events);
-    Mockito.when(blockService.statisticMessage(any(), any())).thenReturn(message);
+    Mockito.when(messageService.statisticMessage(any())).thenReturn(message);
 
     assertDoesNotThrow(() -> {
       publishWeekStatsService.exportStat();
@@ -112,7 +109,7 @@ class PublishWeekStatsServiceTest {
     );
 
     Mockito.when(ghEventService.getEvents(any(), any())).thenReturn(events);
-    Mockito.when(blockService.statisticMessage(any(), any())).thenReturn(message);
+    Mockito.when(messageService.statisticMessage(any())).thenReturn(message);
 
     assertDoesNotThrow(() -> {
       publishWeekStatsService.exportStat();
@@ -158,7 +155,7 @@ class PublishWeekStatsServiceTest {
     );
 
     Mockito.when(ghEventService.getEvents(any(), any())).thenReturn(events);
-    Mockito.when(blockService.statisticMessage(any(), any())).thenReturn(message);
+    Mockito.when(messageService.statisticMessage(any())).thenReturn(message);
 
     assertDoesNotThrow(() -> {
       publishWeekStatsService.exportStat();
